+++ src/core/settings.rs
@@ -1,4 +1,4 @@
-use crate::{Pattern, Theme};
+use crate::{Event, Pattern, PredefinedThemes, Theme};
 
 #[derive(Debug, Clone)]
 pub struct ProgramSettings {
@@ -9,11 +9,49 @@ pub struct ProgramSettings {
 }
 
 #[allow(unused)]
-pub enum Settings {
-    ShowHelpOnError(bool),
+#[derive(Debug, Clone)]
+pub(crate) struct InternalSettings {
+    pub(crate) show_help_on_all_errors: bool,
+    pub(crate) show_help_on_empty_args: bool,
+    pub(crate) enable_command_suggestions: bool,
+    pub(crate) hide_command_aliases: bool,
+    pub(crate) events_to_override: Vec<Event>,
+    pub(crate) separate_options_and_flags: bool,
+    pub(crate) override_all_default_listeners: bool,
+    pub(crate) auto_include_help_subcommand: bool,
+    pub(crate) enable_tree_view_subcommand: bool,
+}
+
+impl Default for InternalSettings {
+    fn default() -> Self {
+        Self {
+            show_help_on_all_errors: false,
+            show_help_on_empty_args: true,
+            enable_command_suggestions: true,
+            hide_command_aliases: true,
+            separate_options_and_flags: false,
+            override_all_default_listeners: false,
+            events_to_override: vec![],
+            auto_include_help_subcommand: true,
+            enable_tree_view_subcommand: true,
+        }
+    }
+}
+
+#[allow(unused)]
+pub enum Setting {
+    ShowHelpOnAllErrors(bool),
+    ShowHelpOnEmptyArgs(bool),
     EnableCommandSuggestion(bool),
     HideCommandAliases(bool),
     SeparateOptionsAndFlags(bool),
+    DefineCustomTheme(Theme),
+    ChoosePredefinedTheme(PredefinedThemes),
+    SetProgramPattern(Pattern),
+    OverrideAllDefaultListeners(bool),
+    OverrideSpecificEventListener(Event),
+    AutoIncludeHelpSubcommand(bool),
+    EnableTreeViewSubcommand(bool),
 }
 
 impl ProgramSettings {