+++ src/parse/matches.rs
@@ -13,7 +13,8 @@ pub struct ParserMatches<'pm> {
     pub(crate) flag_matches: Vec<FlagsMatches<'pm>>,
     pub(crate) option_matches: Vec<OptionsMatches<'pm>>,
     pub(crate) arg_matches: Vec<ArgsMatches>,
-    pub(crate) positional_options: &'pm [&'pm str],
+    pub(crate) positional_args: Vec<String>,
+    pub(crate) marked_args: Vec<(String, i32)>,
 }
 
 #[derive(PartialEq, Eq, Debug, Clone)]
@@ -39,6 +40,10 @@ impl<'o> OptionsMatches<'o> {
             cursor_index: 0,
             option: NewOption::default(),
         }
+    }
+
+    pub(crate) fn contains_option(&self, option: &str) -> bool {
+        self.option.long_version == option || self.option.short_version == option
     }
 }
 
@@ -77,7 +82,8 @@ impl<'a> ParserMatches<'a> {
             matched_subcmd: None,
             arg_matches: vec![],
             option_matches: vec![],
-            positional_options: &[],
+            positional_args: vec![],
+            marked_args: vec![],
         }
     }
 
@@ -89,8 +95,38 @@ impl<'a> ParserMatches<'a> {
         self.matched_subcmd
     }
 
-    pub fn get_arg_count(&self) -> usize {
+    pub fn get_raw_args(&self) -> Vec<String> {
+        let mut args = vec![];
+
+        for arg in self.arg_matches.iter() {
+            args.push(arg.raw_value.clone());
+        }
+
+        args
+    }
+
+    pub fn get_arg(&self, val: &str) -> Option<String> {
+        self.arg_matches
+            .iter()
+            .find(|arg| arg.instance_of == val)
+            .map(|a| a.raw_value.clone())
+    }
+
+    pub fn get_raw_args_count(&self) -> usize {
         self.arg_count
+    }
+
+    pub fn get_instances_of(&self, val: &str) -> Vec<&str> {
+        let mut instances = vec![];
+        for opt_cfg in &self.option_matches {
+            for arg_cfg in &opt_cfg.args {
+                if (arg_cfg.instance_of).as_str() == val {
+                    instances.push((arg_cfg.raw_value).as_str())
+                }
+            }
+        }
+
+        instances
     }
 
     pub fn get_flag(&self, val: &str) -> Option<NewFlag> {