+++ src/parse/parser.rs
@@ -2,7 +2,7 @@
 
 use std::collections::HashMap;
 
-use crate::core::errors::CmderError;
+use crate::core::errors::{CmderError, CmderResult};
 use crate::core::new_program::Command;
 use crate::core::EventConfig;
 use crate::Event;
@@ -178,161 +178,250 @@ impl<'a> Parser<'a> {
     }
 }
 
-pub struct NewParser {}
+pub struct NewParser<'a> {
+    cmd: &'a Command<'a>,
+    cursor_index: usize,
+    marked_args: Vec<(String, bool)>,
+    valid_arg_found: bool,
+    allow_trailing_values: bool,
+    parser_cfg: ParserMatches<'a>,
+}
 
-impl NewParser {
+impl<'p> NewParser<'p> {
+    pub fn new(cmd: &'p Command<'p>) -> Self {
+        Self {
+            cmd,
+            cursor_index: 0,
+            marked_args: vec![],
+            valid_arg_found: false,
+            allow_trailing_values: false,
+            parser_cfg: ParserMatches::new(0, cmd),
+        }
+    }
+
+    pub fn prse(&mut self, args: Vec<String>) {}
+
     pub fn parse<'a>(
-        root: &'a Command<'a>,
-        raw_args: &'a [&'a str],
+        cmd: &'a Command<'a>,
+        args: Vec<String>,
         root_cfg: Option<ParserMatches<'a>>,
-    ) -> Result<ParserMatches<'a>, CmderError<'a>> {
-        // Check if args empty
-        if raw_args.is_empty() {
-            // handle empty args
-            // return Ok(ParserMatches::new(0, root));
+    ) -> CmderResult<ParserMatches<'a>> {
+        if args.is_empty() {
+            let arg_name = if !cmd.get_arguments().is_empty() {
+                (&cmd.get_arguments()[0]).clone().literal
+            } else {
+                "<SUB-COMMAND>".to_string()
+            };
+
+            return Err(CmderError::MissingArgument(vec![arg_name]));
         }
 
+        let mut marked_args = vec![];
         let mut config = if let Some(cfg) = root_cfg {
+            marked_args = cfg.marked_args.clone();
             cfg
         } else {
-            ParserMatches::new(raw_args.len(), root)
+            ParserMatches::new(args.len(), cmd)
         };
 
-        // ["image", "ls", "-p", "80", "--", "-xc", "-pv"]
-        for (cursor_idx, arg) in raw_args.iter().enumerate() {
-            if let Some(flag) = resolve_new_flag(root.get_flags(), arg) {
-                // handle flags input
-                let count = config.get_flag_count(flag.long_version);
+        for a in &args {
+            if !marked_args.iter().map(|f| f.0.clone()).any(|x| x == *a) {
+                marked_args.push((a.to_owned(), 0))
+            }
+        }
 
-                let cfg = FlagsMatches {
-                    appearance_count: (count + 1) as usize,
-                    cursor_index: cursor_idx + config.cursor_offset,
-                    flag,
-                };
+        for (cursor_index, arg) in args.iter().enumerate() {
+            if !arg.is_empty() {
+                // if arg.starts_with('-') {
+                //     // It is either a flag, an option, or '--', or unknown option/flag
+                //     if let Some(flag) = resolve_new_flag(cmd.get_flags(), arg.clone()) {
+                //         // parse flag
+                //     } else if let Some(opt) = resolve_new_option(cmd.get_options(), arg.clone()) {
+                //         // parse option
+                //     } else if arg.contains('=') {
+                //         let parts = arg.split('=').collect::<Vec<_>>();
 
-                if !config.contains_flag(cfg.flag.long_version) {
-                    config.flag_matches.push(cfg);
-                }
-            } else if let Some(opt) = resolve_new_option(root.get_options(), arg) {
-                //handle opts input
-                let max_args_len = opt.arguments.len();
-                let cleaned_name = opt.long_version.replace("--", "").replace('-', "_");
-                let count = config.get_option_count(opt.long_version) as usize;
+                //         if let Some(opt) = resolve_new_option(cmd.get_options(), parts[0].into()) {
+                //             // parse option using parts[1]
+                //         }
+                //     } else {
+                //         return Err(CmderError::UnknownOption(arg.clone()));
+                //     }
+                // }
+                // Check whether its a flag
+                if let Some(flag) = resolve_new_flag(cmd.get_flags(), arg.clone()) {
+                    let count = config.get_flag_count(flag.long_version);
 
-                let mut cfg = OptionsMatches {
-                    cursor_index: cursor_idx + config.cursor_offset,
-                    appearance_count: count + 1,
-                    option: opt.clone(),
-                    args: vec![],
-                };
-
-                for (arg_idx, arg) in opt.arguments.into_iter().enumerate() {
-                    let step = arg_idx + 1;
-                    let mut raw_value = String::new();
-
-                    if arg.variadic {
-                        for (index, a) in raw_args.iter().enumerate() {
-                            if index >= cursor_idx && !a.starts_with('-') {
-                                raw_value.push_str(a);
-                                raw_value.push(' ')
-                            }
-                        }
-                    } else if arg_idx <= max_args_len {
-                        // try to any args input values
-                        match raw_args.get(cursor_idx + step) {
-                            Some(v) => {
-                                raw_value.push_str(v);
-                            }
-                            None => {
-                                if arg.required {
-                                    let mut vals =
-                                        vec![arg.literal, raw_args[cursor_idx].to_string()];
-
-                                    return Err(CmderError::OptionMissingArgument(vals));
-                                }
-                            }
-                        }
-                    }
-
-                    let temp_cfg = ArgsMatches {
-                        cursor_index: (cursor_idx + 1) + config.cursor_offset,
-                        instance_of: arg.literal.clone(),
-                        raw_value,
+                    let flag_cfg = FlagsMatches {
+                        appearance_count: (count + 1) as usize,
+                        cursor_index: cursor_index + config.cursor_offset,
+                        flag,
                     };
 
-                    cfg.args.push(temp_cfg);
-                }
-
-                config.option_matches.push(cfg);
-            } else if let Some(sub_cmd) = root.find_subcommand(arg) {
-                // it is a subcommand/command
-                let arguments = sub_cmd.get_arguments();
-                let max_args_len = arguments.len();
-
-                for (arg_idx, arg) in sub_cmd.get_arguments().iter().enumerate() {
-                    let step = arg_idx + 1;
-                    let mut raw_value = String::new();
-
-                    if arg.variadic {
-                        for (index, a) in raw_args.iter().enumerate() {
-                            if index >= cursor_idx && !a.starts_with('-') {
-                                raw_value.push_str(a);
-                                raw_value.push(' ')
-                            }
-                        }
-                    } else if arg_idx <= max_args_len {
-                        // try to any args input values
-                        match raw_args.get(cursor_idx + step) {
-                            Some(v) => {
-                                if v.starts_with('-') {
-                                    let mut vals = vec![arg.literal.clone()];
-                                    return Err(CmderError::MissingArgument(vals));
-                                } else {
-                                    raw_value.push_str(v);
-                                }
-                            }
-                            None => {
-                                if arg.required {
-                                    let mut vals = vec![arg.literal.clone()];
-                                    return Err(CmderError::MissingArgument(vals));
-                                }
-                            }
-                        }
+                    if !config.contains_flag(flag_cfg.flag.long_version) {
+                        config.flag_matches.push(flag_cfg);
                     }
 
-                    let temp_cfg = ArgsMatches {
-                        cursor_index: (cursor_idx + 1) + config.cursor_offset,
-                        instance_of: arg.literal.clone(),
-                        raw_value,
-                    };
+                    // Mark flag as used
+                    marked_args.get_mut(cursor_index).unwrap().1 = 1;
+                    config.marked_args.push((arg.clone(), 1));
+                } else if let Some(opt) = resolve_new_option(cmd.get_options(), arg.clone()) {
+                    let max_args_len = opt.arguments.len();
+                    let cleaned_name = opt.long_version.replace("--", "").replace('-', "_");
+                    let count = config.get_option_count(opt.long_version) as usize;
 
-                    config.arg_matches.push(temp_cfg);
+                    for (opt_arg_idx, opt_arg_val) in opt.arguments.iter().enumerate() {
+                        let opt_arg_val = opt_arg_val.clone();
+                        let step = opt_arg_idx + 1;
+                        let mut raw_arg_value = String::new();
+
+                        if opt_arg_val.variadic {
+                            for (i, a) in args.iter().enumerate() {
+                                if i >= cursor_index && !a.starts_with('-') {
+                                    raw_arg_value.push_str(a);
+                                    raw_arg_value.push(' ')
+                                }
+                            }
+                        } else if opt_arg_idx <= max_args_len {
+                            match args.get(cursor_index + step) {
+                                Some(val) => {
+                                    // Mark value as visited
+                                    marked_args.get_mut(cursor_index + step).unwrap().1 = 1;
+                                    config.marked_args.push((val.clone(), 1));
+
+                                    raw_arg_value.push_str(val.as_str());
+                                }
+                                None => {
+                                    if opt_arg_val.required {
+                                        let values = vec![
+                                            opt_arg_val.literal,
+                                            args[cursor_index].to_string(),
+                                        ];
+                                        return Err(CmderError::OptionMissingArgument(values));
+                                    }
+                                }
+                            }
+                        }
+
+                        let arg_cfg = ArgsMatches {
+                            cursor_index: (cursor_index + 1) + config.cursor_offset,
+                            instance_of: opt_arg_val.literal.clone(),
+                            raw_value: raw_arg_value,
+                        };
+
+                        if config.contains_option(opt.clone().long_version) {
+                            for opt_match in config.option_matches.iter_mut() {
+                                if opt_match.option.long_version == opt.clone().long_version {
+                                    opt_match.args.push(arg_cfg.clone());
+                                    opt_match.appearance_count += 1;
+                                }
+                            }
+                        } else {
+                            let opt_cfg = OptionsMatches {
+                                appearance_count: (count + 1) as usize,
+                                cursor_index: cursor_index + config.cursor_offset,
+                                option: opt.clone(),
+                                args: vec![arg_cfg],
+                            };
+
+                            config.option_matches.push(opt_cfg);
+                        }
+
+                        // Mark option as used
+                        marked_args.get_mut(cursor_index).unwrap().1 = 1;
+                        config.marked_args.push((arg.clone(), 1));
+                    } // end for
+                } else if let Some(sub_cmd) = cmd.find_subcommand(&arg.clone()) {
+                    let arguments = sub_cmd.get_arguments();
+                    let max_args_len = arguments.len();
+
+                    for (cmd_arg_idx, cmd_arg_val) in arguments.iter().enumerate() {
+                        let step = cmd_arg_idx + 1;
+                        let mut raw_value = String::new();
+
+                        if cmd_arg_val.variadic {
+                            for (i, a) in args.iter().enumerate() {
+                                if i >= cursor_index && !a.starts_with('-') {
+                                    raw_value.push_str(a.as_str());
+                                    raw_value.push(' ')
+                                }
+                            }
+                        } else if cmd_arg_idx <= max_args_len {
+                            match args.get(cursor_index + step) {
+                                Some(val) => {
+                                    if val.starts_with('-') {
+                                        // TODO: Create arg and add it to config
+                                        // let mut vals = vec![cmd_arg_val.literal.clone()];
+                                        // return Err(CmderError::MissingArgument(vals));
+                                        continue;
+                                    } else {
+                                        // Mark value as visited
+                                        marked_args.get_mut(cursor_index + step).unwrap().1 = 1;
+                                        config.marked_args.push((val.clone(), 1));
+
+                                        raw_value.push_str(val);
+                                    }
+                                }
+                                None => {
+                                    if cmd_arg_val.required {
+                                        let vals = vec![cmd_arg_val.literal.clone()];
+                                        return Err(CmderError::MissingArgument(vals));
+                                    }
+                                }
+                            }
+                        }
+
+                        let arg_cfg = ArgsMatches {
+                            cursor_index: (cursor_index + 1) + config.cursor_offset,
+                            instance_of: cmd_arg_val.literal.clone(),
+                            raw_value,
+                        };
+
+                        config.arg_matches.push(arg_cfg);
+                    }
+
+                    config.matched_subcmd = Some(sub_cmd);
+                    config.cursor_offset += 1;
+
+                    // Mark subcommand as used
+                    marked_args.get_mut(cursor_index).unwrap().1 = 1;
+                    config.marked_args.push((arg.clone(), 1));
+
+                    // Parse subcommand
+                    return NewParser::parse(
+                        sub_cmd,
+                        args[cursor_index + 1..].to_vec(),
+                        Some(config),
+                    );
+                } else if arg == "--" {
+                    config.marked_args.push((arg.clone(), 1));
+                    for (i, a) in args[cursor_index + 1..].to_vec().iter().enumerate() {
+                        // Mark argument as used
+                        marked_args.get_mut(cursor_index + (i + 1)).unwrap().1 = 1;
+                        config.marked_args.push((a.clone(), 1));
+                        config.positional_args.push(a.to_string());
+                    }
+                    break;
+                } else if marked_args.get(cursor_index).unwrap().1 == 0 {
+                    // Check whether arg is marked as used
+
+                    if arg.starts_with('-') && arg.as_str() != "--" {
+                        return Err(CmderError::UnknownOption(arg.clone()));
+                    }
+
+                    return Err(CmderError::UnknownCommand(arg.clone()));
                 }
-
-                config.cursor_offset += 1;
-                config.matched_subcmd = Some(sub_cmd);
-
-                return NewParser::parse(sub_cmd, &raw_args[1..], Some(config));
-            } else if *arg == "--" {
-                // register positional_options and break loop
-                config.positional_options = &raw_args[(cursor_idx + 1)..];
-                break;
-            } else {
-                // it is either an argument or unknown
-                let max_len = root.get_arguments().len();
-
-                if max_len > 0 && cursor_idx <= max_len && !arg.starts_with('-') {
-                    continue;
-                }
-
-                if arg.starts_with("-") && *arg != "--" {
-                    return Err(CmderError::UnknownOption(arg));
-                }
-
-                return Err(CmderError::UnknownCommand(arg));
             }
         }
 
         Ok(config)
     }
+
+    fn parse_option() {}
+
+    fn parse_flag() {}
+
+    fn parse_subcmd() {}
+
+    fn parse_arg() {}
 }